Move `h.views.api_annotations` -> `h.views.api.annotations`
